Split the declaration for decision variables out into its own header.
