+++ drake/solvers/CMakeLists.txt
@@ -62,7 +62,7 @@ endif()
 
 add_library_with_exports(LIB_NAME drakeOptimization SOURCE_FILES ${optimization_files})
 target_link_libraries(drakeOptimization drakePolynomial)
-drake_install_headers(MathematicalProgram.h Optimization.h system_identification.h Constraint.h solution_result.h)
+drake_install_headers(decision_variable.h MathematicalProgram.h Optimization.h system_identification.h Constraint.h solution_result.h)
 pods_install_libraries(drakeOptimization)
 pods_install_pkg_config_file(drake-optimization
   LIBS -ldrakeOptimization
@@ -117,4 +117,3 @@ add_subdirectory(test)
 add_subdirectory(+drakeFunction)
 add_subdirectory(BMI)
 add_subdirectory(trajectoryOptimization)
-
